+++ src/fragments/forms/fields-container/components/form-fields/form-fields.js
@@ -53,7 +53,7 @@ export default {
     },
     isSideBarOpen: {
       get () {
-        return this.$store.getters.leftSideBarOpen && !this.$store.getters.embed
+        return this.$store.getters.isSidebarVisible && !this.$store.getters.embed
       },
       set (open) {
         this.$store.commit('setLeftSideBarIsOpen', open)