+++ src/fragments/forms/map-form/components/print/components/isochrones-template/PrintIsochrones.vue
@@ -1,0 +1,26 @@
+<template>
+  <div class="priting-template directions-template">
+    <h2 v-html="contentTitle"></h2>
+    <br /><br />
+
+    <img width="auto" height="400" :src="mapViewImage" />
+    <template v-for="(place, index) in localMapViewData.places">
+      <div :key="index">
+        <h5>{{place.placeName}}</h5>          
+        <template v-for="(polygon, polygonIndex) in localMapViewData.polygons">
+          <template v-if="hasAsCenter(place, polygon)">
+            <div :key="polygonIndex" style="padding-left:10px">
+              <span :style="{background: polygon.properties.color}">
+                <b :style="{color: polygonAreaTextColor(polygon)}">{{polygon.properties.label.replace('Polygon', '')}}</b> 
+              </span>
+              <span>&rArr;</span> {{polygon.properties.area}}
+            </div>
+          </template>
+        </template>
+        <br />
+       </div>
+    </template>
+  </div>
+</template>
+
+<script src="./print-isochrones.js"></script>