+++ src/fragments/forms/map-form/components/print/components/place-template/print-place.js
@@ -1,0 +1,24 @@
+import MapViewData from '@/models/map-view-data'
+
+export default {
+  props: {
+    mapViewData: {
+      type: MapViewData
+    },
+    mapViewImage: {
+      type: String
+    }
+  },
+  created () {
+    this.localMapViewData = this.mapViewData
+  },
+  data: () => ({
+    localMapViewData: null,
+    mapImageDataUrl: null
+  }),
+  computed: {
+    place () {
+      return this.localMapViewData.places[0]
+    }    
+  }
+}