+++ src/fragments/forms/map-form/components/print/i18n/print.i18n.de-de.js
@@ -1,0 +1,9 @@
+
+export default {
+  print: {
+    print: 'Ausdrucken',
+    segment: 'Abschnitt',
+    preparingPrinting: 'Ausdruck vorbereiten',
+    readyToPrint: 'Bereit zum Drucken'
+  }
+}