+++ src/fragments/forms/map-form/components/print/i18n/print.i18n.fr-fr.js
@@ -1,0 +1,9 @@
+
+export default {
+  print: {
+    print: 'Imprimer',
+    segment: 'Segment',
+    preparingPrinting: 'Préparation de l\'impression',
+    readyToPrint: 'Prêt à imprimer'
+  }
+}