+++ src/fragments/forms/map-form/components/print/i18n/print.i18n.it-it.js
@@ -1,0 +1,9 @@
+
+export default {
+  print: {
+    print: 'Stampare',
+    segment: 'Segmento',
+    preparingPrinting: 'Preparazione della stampa',
+    readyToPrint: 'Pronta per stampare'
+  }
+}